+++ src/main/java/com/kusitms/jipbap/user/UserController.java
@@ -1,5 +1,6 @@
 package com.kusitms.jipbap.user;
 
+import com.kusitms.jipbap.auth.dto.ReissueResponseDto;
 import com.kusitms.jipbap.common.response.CommonResponse;
 import com.kusitms.jipbap.security.Auth;
 import com.kusitms.jipbap.security.AuthInfo;
@@ -17,6 +18,12 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     private final UserService userService;
+
+    @Operation(summary = "액세스 토큰 재발급 - 헤더에 refreshToken 정보 포함하여 요청")
+    @PostMapping("/reissue")
+    public CommonResponse<ReissueResponseDto> reissue(@Auth AuthInfo authInfo) {
+        return new CommonResponse<>(userService.reissue(authInfo.getEmail(), authInfo.getToken()));
+    }
 
     @Operation(summary = "로그아웃 - 리프레쉬 토큰 삭제")
     @PostMapping("/logout")
